Rename index to migrator
